+++ resources/views/shared/sugar/clarified.blade.php
@@ -1,0 +1,9 @@
+{{
+  Form::bsSelect($name, 'Clarified by', [
+    'Join Study' => 'Join Study',
+    'In House Study' => 'In House Study',
+    'Infill Seismic 2D' => 'Infill Seismic 2D',
+    'Seismic 3D' => 'Seismic 3D',
+    'Third Party Study' => 'Third Party Study'
+  ], true)
+}}